requirements.txt
anyio==3.7.0
bcrypt==4.0.1
boto3==1.34.49
botocore==1.34.49
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
distlib==0.3.6
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-session==0.2.7
fastapi-sessions==0.3.2
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
jose==1.0.0
loguru==0.7.2
passlib==1.7.4
platformdirs==3.5.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
python-dateutil==2.8.2
python-jose==3.3.0
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.6
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
urllib3==2.0.7
uvicorn==0.22.0
virtualenv==20.23.0
win32-setctime==1.1.0